feat: Add receiveMovingCards(). Rewrite doMoveAndFlip().
Animated cards are now pulled by a Pile, not pushed by CardView
or CardMoves. That allows Moves to be made instantaneously from
Pile to Pile, leaving the animation to catch up later, and the
Pile to calculate where to put the cards in a fanned out Pile.
Cards can also be seen to flip over while moving, if required.