fix: CardMoves mostly working. Recoded replenishTableau algorithm.
Added the _replaceTableauCard() callback "loop". If an excluded
card has been dealt from the Stock to a Tableau, the procedure is
called ONCE to start the replacement and cites itself as callback.
So it "loops" until a valid card is received. It is all asynch and
fully animated as cards fly out to the rejects pile and in from
the Stock Pile.